홍영현 : 스킬 및 데미지 보완, 보스 체력에 따라 패턴, 인피니트 끝날시 소리 재생

+++ DirecX_Maple/GameEngineContents/BaseSummonActor.h
@@ -60,7 +60,7 @@ protected:
 	float DirAngle = 0.0f;
 	float Speed = 800.0f;
 	float RotationSpeed = 0.0f;
-	size_t WonderHitCount = 3;
+	int WonderHitCount = 3;
 
 
 	WonderState State = WonderState::Start;
